feat(realtime): notify clients on new public/community posts & refactor post form
- Add WebSocket notification system for new posts (public/community only)
- Notify all connected clients except the post creator when a new public or community post is created
- Move PostNotifier logic to a dedicated module for better separation of concerns
- Refactor post creation form into a reusable PostFormModal component (React)
- Show post creation form as a popup modal instead of inline
- Clean up unused variables and code in Home.jsx after refactor
- Auto-format and lint codebase for both backend (Python) and frontend (React)
- Ensure notification only appears for users other than the post creator

BREAKING CHANGE: The post creation UI is now a popup modal. Notification logic is now handled via WebSocket and only triggers for public/community posts and for users other than the creator.
